+++ src/ton/tonInfo.ts
@@ -6,17 +6,10 @@ import { asTonInfoPayload, TonInfoPayload, TonNetworkInfo } from './tonTypes'
 
 const networkInfo: TonNetworkInfo = {
   defaultWalletContract: 'v5r1',
+  drpcUrl: 'https://toncenter.com/api/v2',
   minimumAddressBalance: '50000000', // 0.5 TON There isn't a hardcoded minimum but the user needs to keep something left
   pluginMnemonicKeyName: 'tonMnemonic',
-  tonCenterUrl: 'https://toncenter.com/api/v2/jsonRPC',
-  tonOrbsServers: [
-    'https://ton.access.orbs.network/4410c0ff5Bd3F8B62C092Ab4D238bEE463E64410/1/mainnet/toncenter-api-v2/jsonRPC',
-    'https://ton.access.orbs.network/4411c0ff5Bd3F8B62C092Ab4D238bEE463E64411/1/mainnet/toncenter-api-v2/jsonRPC',
-    'https://ton.access.orbs.network/4412c0ff5Bd3F8B62C092Ab4D238bEE463E64412/1/mainnet/toncenter-api-v2/jsonRPC',
-    'https://ton.access.orbs.network/55013c0ff5Bd3F8B62C092Ab4D238bEE463E5501/1/mainnet/toncenter-api-v2/jsonRPC',
-    'https://ton.access.orbs.network/55023c0ff5Bd3F8B62C092Ab4D238bEE463E5502/1/mainnet/toncenter-api-v2/jsonRPC',
-    'https://ton.access.orbs.network/55033c0ff5Bd3F8B62C092Ab4D238bEE463E5503/1/mainnet/toncenter-api-v2/jsonRPC'
-  ]
+  tonCenterUrl: 'https://toncenter.com/api/v2/jsonRPC'
 }
 
 const currencyInfo: EdgeCurrencyInfo = {